requirements.txt
discord.py
yt-dlp
spotipy
PyNaCl
